MDEV-23326: mysql_tzinfo_to_sql add --skip-wsrep-check
Several mysql_tzinfo_to_sql behaviours are dependent on
checking the wsrep_on==ON specificly for galera initialization
of timezone (or skipping of them).

There are two scenarios where the lack of the checking would
be very useful.

With the --skip-write-binary-log, binary logs are explictly
skipped, allowing for loading in classical replication scenarios.

Without the --skip-write-binary-log, innodb can be used for
loading the timezone tables signficantly faster than Aria
with TRANSACTIONAL=1.

The lack of checking the wsrep status also make it possible
to use mysql_tzinfo_to_sql in the mysqld --bootstrap input
allowing for a single initialization of all system tables.

In cleaning up part of MDEV-18778, when changing back from InnoDB
to MyISAM, include the ORDER BY, and don't include this if only
loading leap seconds.